pyproject.toml
[project]
name = "sorawatermarkcleaner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.21.0",
    "diffusers>=0.35.1",
    "einops>=0.8.1",
    "fastapi==0.108.0",
    "ffmpeg-python>=0.2.0",
    "fire>=0.7.1",
    "google-cloud-storage>=2.10.0",
    "httpx>=0.28.1",
    "huggingface-hub>=0.35.3",
    "jupyter>=1.1.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "notebook>=7.4.7",
    "omegaconf>=2.3.0",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "pydantic>=2.11.10",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "ruptures>=1.1.10",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.50.0",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.0",
    "ultralytics>=8.3.204",
    "uuid>=1.30",
    "uvicorn>=0.35.0",
]

